exceptions

+++ MVC/MyGameStore/MyGameStore.BLL/Services/PersonService.cs
@@ -1,6 +1,7 @@
 ﻿using MyGameStore.BLL.Interfaces;
 using MyGameStore.DAL;
 using MyGameStore.DAL.UOW;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 
@@ -16,12 +17,16 @@ namespace MyGameStore.BLL.Services
         }
         public void Add(Person person)
         {
+            if (_unitOfWork.StoreRepository.GetById((int)person.StoreId) == null)
+                throw new Exception("Store id doesn't exist");
             _unitOfWork.PersonRepository.Create(person);
             _unitOfWork.Commit();
         }
 
         public void Delete(int id)
         {
+            if (GetbyId(id) == null)
+                throw new Exception("Person not found");
             _unitOfWork.PersonRepository.Delete(id);
             _unitOfWork.Commit();
         }
